--- BookPlayer/Library/LibraryViewController.swift
@@ -1,365 +1,0 @@
-//
-//  LibraryViewController.swift
-//  BookPlayer
-//
-//  Created by Gianni Carlo on 7/7/16.
-//  Copyright © 2016 Tortuga Power. All rights reserved.
-//
-
-import BookPlayerKit
-import Combine
-import CoreData
-import MediaPlayer
-import UIKit
-
-// swiftlint:disable file_length
-
-class LibraryViewController: ItemListViewController, UIGestureRecognizerDelegate {
-  private var fileSubscription: AnyCancellable?
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        self.navigationItem.title = "library_title".localized
-        self.navigationItem.leftBarButtonItem?.title = "settings_button".localized
-
-        // VoiceOver
-        self.setupCustomRotors()
-
-        // enables pop gesture on pushed controller
-        self.navigationController!.interactivePopGestureRecognizer!.delegate = self
-
-        self.handleCoreDataError()
-
-        self.loadLibrary()
-
-        self.loadLastBook()
-    }
-
-    // No longer need to deregister observers for iOS 9+!
-    // https://developer.apple.com/library/mac/releasenotes/Foundation/RN-Foundation/index.html#10_11NotificationCenter
-    deinit {
-        // for iOS 8
-        NotificationCenter.default.removeObserver(self)
-        UIApplication.shared.endReceivingRemoteControlEvents()
-    }
-
-    override func setupObservers() {
-      super.setupObservers()
-      // register for appDelegate openUrl notifications
-      NotificationCenter.default.addObserver(self, selector: #selector(self.reloadData), name: .reloadData, object: nil)
-      NotificationCenter.default.addObserver(self, selector: #selector(self.onProcessingFile(_:)), name: .processingFile, object: nil)
-      NotificationCenter.default.addObserver(self, selector: #selector(self.onNewFileUrl), name: .newFileUrl, object: nil)
-      NotificationCenter.default.addObserver(self, selector: #selector(self.onDownloadingProgress(_:)), name: .downloadProgress, object: nil)
-      NotificationCenter.default.addObserver(forName: .reloadLibrary, object: nil, queue: .main) { [weak self] _ in
-        self?.loadLibrary()
-      }
-    }
-
-  func showImport() {
-    self.fileSubscription?.cancel()
-    self.coordinator?.showImport()
-  }
-
-  func handleCoreDataError() {
-    guard let coreDataError = DataManager.loadingDataError else { return }
-
-    let error = coreDataError as NSError
-    // CoreData may fail if device doesn't have space
-    if (error.domain == NSPOSIXErrorDomain && error.code == ENOSPC) ||
-        (error.domain == NSCocoaErrorDomain && error.code == NSFileWriteOutOfSpaceError) {
-      self.showAlert("error_title".localized, message: "coredata_error_diskfull_description".localized)
-      return
-    }
-
-    // Handle data error migration by reloading library
-    if error.code == NSMigrationError ||
-        error.code == NSMigrationConstraintViolationError ||
-        error.code == NSMigrationCancelledError ||
-        error.code == NSMigrationMissingSourceModelError ||
-        error.code == NSMigrationMissingMappingModelError ||
-        error.code == NSMigrationManagerSourceStoreError ||
-        error.code == NSMigrationManagerDestinationStoreError ||
-        error.code == NSEntityMigrationPolicyError {
-      self.showAlert("error_title".localized, message: "coredata_error_migration_description".localized) {
-        DataManager.cleanAndReloadLibrary()
-      }
-      return
-    }
-
-    fatalError("Unresolved error \(error), \(error.userInfo)")
-  }
-
-    func loadLibrary() {
-      guard let library = try? DataManager.getLibrary() ?? DataManager.createLibrary() else { return }
-
-      self.library = library
-
-      self.toggleEmptyStateView()
-
-      self.tableView.reloadData()
-
-      DataManager.notifyPendingFiles()
-    }
-
-    func downloadBook(from urlString: String) {
-        guard let url = URL(string: urlString) else {
-            self.showAlert("error_title".localized, message: String.localizedStringWithFormat("invalid_url_title".localized, urlString))
-            return
-        }
-
-        self.showLoadView(true, title: "downloading_file_title".localized, subtitle: "\("progress_title".localized) 0%")
-
-        NetworkService.shared.download(from: url) { response in
-            self.showLoadView(false)
-
-            if response.error != nil,
-                let error = response.error {
-                self.showAlert("network_error_title".localized, message: error.localizedDescription)
-            }
-
-            if let response = response.response, response.statusCode >= 300 {
-                self.showAlert("network_error_title".localized, message: "Code \(response.statusCode)")
-            }
-        }
-    }
-
-    func loadLastBook() {
-      guard self.library != nil, let book = self.library.lastPlayedBook else {
-        return
-      }
-
-      self.coordinator?.loadLastBook(book)
-    }
-
-    func gestureRecognizerShouldBegin(_: UIGestureRecognizer) -> Bool {
-        return navigationController!.viewControllers.count > 1
-    }
-
-    // MARK: - Callback events
-
-    // This is called from a background thread inside an ImportOperation
-    @objc func onDownloadingProgress(_ notification: Notification) {
-        guard
-            let userInfo = notification.userInfo,
-            let progress = userInfo["progress"] as? String else {
-            return
-        }
-
-        DispatchQueue.main.async {
-            self.showLoadView(true, title: "downloading_file_title".localized, subtitle: "\("progress_title".localized) \(progress)%")
-
-            if let vc = self.navigationController?.visibleViewController as? PlaylistViewController {
-                vc.showLoadView(true, title: "downloading_file_title".localized, subtitle: "\("progress_title".localized) \(progress)%")
-            }
-        }
-    }
-
-    @objc func onNewFileUrl() {
-        guard self.loadingView.isHidden else { return }
-        let loadingTitle = "import_preparing_title".localized
-        UIAccessibility.post(notification: UIAccessibility.Notification.layoutChanged, argument: self.loadingView.titleLabel)
-
-        if let vc = self.navigationController?.visibleViewController as? PlayerViewController {
-          vc.viewModel.dismiss()
-        }
-
-        self.showLoadView(true, title: loadingTitle)
-
-        if let vc = self.navigationController?.visibleViewController as? PlaylistViewController {
-            vc.showLoadView(true, title: loadingTitle)
-        }
-    }
-
-    // This is called from a background thread inside an ImportOperation
-    @objc func onProcessingFile(_ notification: Notification) {
-        guard
-            let userInfo = notification.userInfo,
-            let filename = userInfo["filename"] as? String else {
-            return
-        }
-
-        DispatchQueue.main.async {
-            self.showLoadView(true, title: nil, subtitle: filename)
-
-            if let vc = self.navigationController?.visibleViewController as? PlaylistViewController {
-                vc.showLoadView(true, title: nil, subtitle: filename)
-            }
-        }
-    }
-
-    // MARK: - IBActions
-
-    @IBAction func addAction() {
-        self.presentAddOptionsAlert()
-    }
-
-    // Sorting
-    override func sort(by sortType: PlayListSortOrder) {
-        library.sort(by: sortType)
-    }
-
-    override func presentAddOptionsAlert() {
-        let alertController = UIAlertController(title: nil,
-                                                message: "import_description".localized,
-                                                preferredStyle: .actionSheet)
-
-        alertController.addAction(UIAlertAction(title: "import_button".localized, style: .default) { _ in
-            self.presentImportFilesAlert()
-        })
-
-        alertController.addAction(UIAlertAction(title: "create_playlist_button".localized, style: .default) { _ in
-            self.presentCreateFolderAlert(handler: { title in
-              do {
-                _ = try DataManager.createFolder(with: title, in: nil, library: self.library)
-              } catch {
-                self.showAlert("error_title".localized, message: error.localizedDescription)
-              }
-
-              self.reloadData()
-            })
-        })
-
-        alertController.addAction(UIAlertAction(title: "cancel_button".localized, style: .cancel))
-
-        present(alertController, animated: true, completion: nil)
-    }
-
-    override func handleMove(_ selectedItems: [LibraryItem]) {
-        let alert = UIAlertController(title: "choose_destination_title".localized, message: nil, preferredStyle: .alert)
-
-        alert.addAction(UIAlertAction(title: "new_playlist_button".localized, style: .default) { _ in
-            self.presentCreateFolderAlert(handler: { title in
-              do {
-                let folder = try DataManager.createFolder(with: title, in: nil, library: self.library)
-                try self.move(selectedItems, to: folder)
-              } catch {
-                self.showAlert("error_title".localized, message: error.localizedDescription)
-              }
-            })
-        })
-
-        let availableFolders = self.items.compactMap { (item) -> Folder? in
-            item as? Folder
-        }
-
-        let existingFolderAction = UIAlertAction(title: "existing_playlist_button".localized, style: .default) { _ in
-        }
-
-        existingFolderAction.isEnabled = !availableFolders.isEmpty
-        alert.addAction(existingFolderAction)
-
-        alert.addAction(UIAlertAction(title: "cancel_button".localized, style: .cancel))
-
-        self.present(alert, animated: true, completion: nil)
-    }
-
-    override func handleTrash(_ selectedItems: [LibraryItem]) {
-        self.handleDelete(items: selectedItems)
-    }
-}
-
-// MARK: Accessibility
-
-extension LibraryViewController {
-    private func setupCustomRotors() {
-        accessibilityCustomRotors = [self.rotorFactory(name: "Books", type: .book), self.rotorFactory(name: "Folders", type: .folder)]
-    }
-
-    private func rotorFactory(name: String, type: SimpleItemType) -> UIAccessibilityCustomRotor {
-        return UIAccessibilityCustomRotor(name: name) { (predicate) -> UIAccessibilityCustomRotorItemResult? in
-            let forward: Bool = (predicate.searchDirection == .next)
-
-            let playListCells = self.tableView.visibleCells.filter { (cell) -> Bool in
-                guard let cell = cell as? BookCellView else { return false }
-                return cell.type == type
-            }
-
-            var currentIndex = forward ? -1 : playListCells.count
-            //
-            if let currentElement = predicate.currentItem.targetElement {
-                if let cell = currentElement as? BookCellView {
-                    currentIndex = playListCells.firstIndex(of: cell) ?? currentIndex
-                }
-            }
-            let nextIndex = forward ? currentIndex + 1 : currentIndex - 1
-
-            while nextIndex >= 0, nextIndex < playListCells.count {
-                let cell = playListCells[nextIndex]
-                return UIAccessibilityCustomRotorItemResult(targetElement: cell, targetRange: nil)
-            }
-            return nil
-        }
-    }
-}
-
-// MARK: - TableView Delegate
-
-extension LibraryViewController {
-  func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
-    guard indexPath.sectionValue == .data else { return nil }
-
-    let item = items[indexPath.row]
-
-    let optionsAction = UIContextualAction(style: .normal, title: "\("options_button".localized)…") { _, _, completion in
-      guard let sheet = self.createOptionsSheetController([item]) else { return }
-
-      // "…" on a button indicates a follow up dialog instead of an immmediate action in macOS and iOS
-      var title = "\("delete_button".localized)…"
-
-      // Remove the dots if trying to delete an empty folder
-      if let folder = item as? Folder {
-          title = folder.hasBooks() ? title : "delete_button".localized
-      }
-
-      let deleteAction = UIAlertAction(title: title, style: .destructive) { _ in
-          guard let book = self.items[indexPath.row] as? Book else {
-              guard let folder = self.items[indexPath.row] as? Folder else { return }
-
-              guard folder.hasBooks() else {
-                do {
-                  try DataManager.delete([folder], library: self.library)
-                  self.deleteRows(at: [indexPath])
-                } catch {
-                  self.showAlert("error_title".localized, message: error.localizedDescription)
-                }
-
-                return
-              }
-
-              self.handleDelete(items: [folder])
-
-              return
-          }
-
-          self.handleDelete(items: [book])
-      }
-
-      sheet.addAction(deleteAction)
-
-      self.present(sheet, animated: true, completion: nil)
-      completion(true)
-    }
-
-    return UISwipeActionsConfiguration(actions: [optionsAction])
-  }
-}
-
-// MARK: - TableView DataSource
-
-extension LibraryViewController {
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = super.tableView(tableView, cellForRowAt: indexPath)
-
-        guard let bookCell = cell as? BookCellView,
-            let currentBook = PlayerManager.shared.currentBook,
-            let index = self.library.index(for: currentBook),
-            index == indexPath.row else {
-            return cell
-        }
-
-        bookCell.playbackState = .paused
-
-        return bookCell
-    }
-}
